backend/requirements.txt
flask
supabase-py
gunicorn
python-dotenv
flask-cors